Fixed timestamp not uint64_t

+++ commons/ipc.h
@@ -19,7 +19,7 @@ union P_PACKET {
 struct B_INIT {
 	bool astronomical;
 	double mass;
-	unsigned int time;
+	uint64_t time;
 	Eigen::Vector3d pos;
 	Eigen::Vector3d vel;
 };
